Move PhotoCloudService
