+++ tests/api/Services/PhotoCloudServiceTest.cs
@@ -3,11 +3,10 @@ using BDMS.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using System.Security.Claims;
 
-namespace BDMS;
+namespace BDMS.Services;
 
 [TestClass]
 public class PhotoCloudServiceTest
